moved pages to pages and components to components
