+++ client/src/App.js
@@ -4,27 +4,22 @@ import { setContext } from '@apollo/client/link/context';
 
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
-// import SearchGames from "./pages/SearchGames";
 import NotFound from './pages/NotFound';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
-import AltRules from './pages/AltRules';
-import Homepage from "./components/Homepage";
-import Play from "./pages/Play"; 
-import Results from './pages/Results'; 
+import AltRules from './components/AltRules';
+import Homepage from "./pages/Homepage";
 import Friends from './pages/Friends';
+import SearchGames from "./pages/SearchGames";
 
-import SearchGames from "./components/SearchGames";
 // import GameSearch from "./components/GameSearch";
 import { ThemeProvider } from "@mui/material/styles";
 import theme from "./Theme/Theme";
-// import SubmitBtn from "./components/SubmitBtn";
-// import PieChartDemo from "./pages/PieChartDemo";
 
-import FunFact from "./pages/FactArray";
+import FunFact from "./components/FactArray";
 // import AltRules from "./pages/AltRules";
 // import { withRouter } from 'react-router-dom';
-import Profile from "./components/Profile"
+import Profile from "./pages/Profile"
 
 import {
   ApolloClient,
@@ -72,8 +67,6 @@ function App() {
 								<Route path="/signup" element={<Signup/>} />
 								<Route path="/profile" element={<Profile/>}/>
 								<Route path="/altrules" element={<AltRules/>}/>
-								<Route path='/play' element={<Play/>}/>
-								<Route path='/results' element={<Results/>}/>
 								<Route path="/friends" element={<Friends/>}/>
 								{/* <Route path="/startmatch" element={<StartMatch/>}/> */}
 								<Route path="*" element={<NotFound/>}/>
