+++ client/src/components/Drawer.js
@@ -9,7 +9,6 @@ import {
   makeStyles,
 } from "@material-ui/core";
 import { Link } from "react-router-dom";
-
 import MenuIcon from "@mui/icons-material/Menu";
 
 const useStyles = makeStyles(()=>({
